+++ ZPCIL_test/SCT.h
@@ -1,10 +1,10 @@
 // Синус, косинус, тангенс угла А
 #include <math.h>
 
-void SCT(float A, float &SNA, float &CSA, float &TGA)
+void SCT(double A, float &SNA, float &CSA, float &TGA)
 {
-	SNA = sin(A);
-	CSA = cos(A);
-	if (A == 0.) TGA = 1.E-6;
+	SNA = static_cast<float>(sin(A));
+	CSA = static_cast<float>(cos(A));
+	if (A == 0.) TGA = static_cast<float>(1.E-6);
 	else TGA = SNA / CSA;	
 }
